requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
sentencepiece
evaluate
Pillow
opencv-python
flask
pytest